处理登录保存用户信息

+++ app/src/main/java/com/wkk/wanandroid/dao/UserDao.kt
@@ -30,5 +30,5 @@ interface UserDao {
     suspend fun delete(user: User)
 
     @Query("SELECT * FROM users WHERE id=:userId")
-    suspend fun getUserById(userId: Int): User
+    suspend fun getUserById(userId: Long): User
 }
